Extract some inner classes out and address review comments.
