+++ flink-state-backends/flink-statebackend-heap-spillable/src/main/java/org/apache/flink/runtime/state/heap/space/PowerTwoBucketAllocator.java
@@ -22,11 +22,7 @@ import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.util.MathUtils;
 import org.apache.flink.util.Preconditions;
 
-import org.apache.commons.collections.map.LinkedMap;
-
 import java.util.concurrent.ConcurrentLinkedQueue;
-
-import static org.apache.flink.runtime.state.heap.space.SpaceConstants.NO_SPACE;
 
 /**
  * This is a simple buddy-like allocator used for small space. The allocator always allocates a block
@@ -157,262 +153,4 @@ public class PowerTwoBucketAllocator implements BucketAllocator {
 		return 1 << (blockAllocatorIndex + MIN_BLOCK_SIZE_BITS);
 	}
 
-	/**
-	 * A bucket is a continuous space that is divided into blocks with fixed size.
-	 * Buckets can be reused to allocate another size of blocks after it's freed.
-	 */
-	public static final class Bucket {
-
-		/**
-		 * Base offset of this bucket in the chunk.
-		 */
-		private final int baseOffset;
-
-		/**
-		 * Size of this bucket.
-		 */
-		private final int bucketSize;
-
-		/**
-		 * Index of block allocator which is using this bucket. -1 indicates
-		 * this bucket is not used by any allocator.
-		 */
-		private int blockAllocatorIndex;
-
-		/**
-		 * Size of block to allocate. This field is only meaningful when the
-		 * bucket is used by some allocator.
-		 */
-		private int blockSize;
-
-		/**
-		 * Number of bits to represent the block size.
-		 */
-		private int blockSizeBits;
-
-		/**
-		 * Number of free blocks in bucket.
-		 */
-		private int freeBlockNum;
-
-		/**
-		 * Number of used blocks in bucket.
-		 */
-		private int usedBlockNum;
-
-		/**
-		 * An array of indexes for free blocks. Blocks in [0, getFreeBlockNum)
-		 * are free to use. This works like a stack. For allocate, last block
-		 * is popped for the back of array, and for free, block are pushed to
-		 * the back of array.
-		 */
-		private int[] freeBlockIndexes;
-
-		Bucket(int offset, int bucketSize) {
-			this.baseOffset = offset;
-			this.bucketSize = bucketSize;
-			this.blockAllocatorIndex = -1;
-		}
-
-		/**
-		 * Reuse the bucket by the given allocator.
-		 *
-		 * @param bucketAllocatorIndex index of bucket allocator.
-		 */
-		void reuseBucket(int bucketAllocatorIndex) {
-			this.blockAllocatorIndex = bucketAllocatorIndex;
-			this.blockSize = getBlockSizeFromBlockAllocatorIndex(this.blockAllocatorIndex);
-			Preconditions.checkArgument(this.blockSize <= this.bucketSize,
-				"Block size can't be larger than bucket size.");
-			this.blockSizeBits = MathUtils.log2strict(this.blockSize);
-
-			int totalBlockNum = this.bucketSize >>> this.blockSizeBits;
-			this.freeBlockNum = totalBlockNum;
-			this.usedBlockNum = 0;
-			this.freeBlockIndexes = new int[totalBlockNum];
-			for (int i = 0; i < totalBlockNum; ++i) {
-				this.freeBlockIndexes[i] = i;
-			}
-		}
-
-		/**
-		 * Allocates a block.
-		 *
-		 * @return offset of the block.
-		 */
-		public int allocate() {
-			// pop the last block
-			++usedBlockNum;
-			--freeBlockNum;
-			return baseOffset + (freeBlockIndexes[freeBlockNum] << blockSizeBits);
-		}
-
-		/**
-		 * Frees the block with the given offset.
-		 *
-		 * @param offset offset of the block to free.
-		 */
-		public void free(int offset) {
-			int bucketOffset = offset - baseOffset;
-			int blockIndex = bucketOffset >>> this.blockSizeBits;
-			--usedBlockNum;
-			// push the block to the back
-			freeBlockIndexes[freeBlockNum++] = blockIndex;
-		}
-
-		/**
-		 * Returns true if this bucket has free blocks to allocate.
-		 */
-		boolean hasFreeBlocks() {
-			return freeBlockNum > 0;
-		}
-
-		/**
-		 * Returns true if there is no blocks used in this bucket.
-		 */
-		public boolean isFree() {
-			return usedBlockNum == 0;
-		}
-
-		int getBlockAllocatorIndex() {
-			return blockAllocatorIndex;
-		}
-
-		int getBlockSize() {
-			return blockSize;
-		}
-
-		boolean isUninitiated() {
-			return blockAllocatorIndex == -1;
-		}
-
-		int getFreeBlockNum() {
-			return freeBlockNum;
-		}
-
-		int getUsedBlockNum() {
-			return usedBlockNum;
-		}
-
-		int getFreeBytes() {
-			return freeBlockNum << blockSizeBits;
-		}
-
-		int getUsedBytes() {
-			return usedBlockNum << blockSizeBits;
-		}
-
-		long getBaseOffset() {
-			return baseOffset;
-		}
-	}
-
-	/**
-	 * Allocates blocks with fixed size from buckets.
-	 */
-	static final class BlockAllocator {
-
-		/**
-		 * Index of this block allocator.
-		 */
-		private final int blockAllocatorIndex;
-
-		/**
-		 * Queue of free buckets. Allocator can apply a free bucket
-		 * from the queue, and returns a free bucket to the queue.
-		 */
-		private final ConcurrentLinkedQueue<Bucket> freeBucketsQueue;
-
-		/**
-		 * Map of buckets applied by this allocator from queue.
-		 */
-		private LinkedMap appliedBuckets;
-
-		/**
-		 * Map of buckets that have free blocks to allocate.
-		 */
-		private LinkedMap freeBuckets;
-
-		BlockAllocator(int blockAllocatorIndex, ConcurrentLinkedQueue<Bucket> freeBucketsQueue) {
-			this.blockAllocatorIndex = blockAllocatorIndex;
-			this.freeBucketsQueue = freeBucketsQueue;
-			this.appliedBuckets = new LinkedMap();
-			this.freeBuckets = new LinkedMap();
-		}
-
-		/**
-		 * Allocates a block.
-		 *
-		 * @return offset of the block.
-		 */
-		synchronized int allocateBlock() {
-			Bucket bucket = null;
-
-			// 1. try to find a bucket that have free blocks
-			if (freeBuckets.size() > 0) {
-				bucket = (Bucket) freeBuckets.lastKey();
-			}
-
-			// 2. try to apply a new bucket from queue
-			if (bucket == null) {
-				bucket = freeBucketsQueue.poll();
-				if (bucket != null) {
-					addBucket(bucket);
-				}
-			}
-
-			if (bucket == null) {
-				return NO_SPACE;
-			}
-
-			int offset = bucket.allocate();
-			if (!bucket.hasFreeBlocks()) {
-				freeBuckets.remove(bucket);
-			}
-
-			return offset;
-		}
-
-		/**
-		 * Frees a block from the given bucket.
-		 *
-		 * @param bucket bucket that block belongs to.
-		 * @param offset offset of block to free.
-		 */
-		synchronized void freeBlock(Bucket bucket, int offset) {
-			bucket.free(offset);
-			if (bucket.isFree()) {
-				removeBucket(bucket);
-				freeBucketsQueue.offer(bucket);
-			} else if (!freeBuckets.containsKey(bucket)) {
-				freeBuckets.put(bucket, bucket);
-			}
-		}
-
-		private void addBucket(Bucket b) {
-			b.reuseBucket(blockAllocatorIndex);
-			appliedBuckets.put(b, b);
-			freeBuckets.put(b, b);
-		}
-
-		private void removeBucket(Bucket b) {
-			appliedBuckets.remove(b);
-			freeBuckets.remove(b);
-		}
-
-		@VisibleForTesting
-		int getBlockAllocatorIndex() {
-			return blockAllocatorIndex;
-		}
-
-		@VisibleForTesting
-		LinkedMap getAppliedBuckets() {
-			return appliedBuckets;
-		}
-
-		@VisibleForTesting
-		LinkedMap getFreeBuckets() {
-			return freeBuckets;
-		}
-	}
 }
