+++ flink-state-backends/flink-statebackend-heap-spillable/src/main/java/org/apache/flink/runtime/state/heap/space/SpaceConstants.java
@@ -23,6 +23,9 @@ package org.apache.flink.runtime.state.heap.space;
  */
 public class SpaceConstants {
 
+	private SpaceConstants() {
+	}
+
 	/** This indicates there is no space left. */
 	static final int NO_SPACE = -1;
